# Add more indexes to speed up queries and reduce load on the database server