--- a/src/clj/tasks/db.clj
+++ b/src/clj/tasks/db.clj
@@ -76,20 +76,41 @@
                          (.printStackTrace e)))
     (finally (webdb/disconnect))))
 
+(def ^:const indexes
+  (let [case-insensitive-index-opts {:collation {:locale "en" :strength (int 2)}}]
+    [["cards" (array-map :code 1)]
+     ["cards" (array-map :previous-versions 1)]
+     ["cards" (array-map :type 1)]
+     ["decks" (array-map :username 1)]
+     ["game-logs" (array-map :gameid 1)]
+     ["game-logs" (array-map "corp.player.username" 1)]
+     ["game-logs" (array-map "runner.player.username" 1)]
+     ["messages" (array-map :channel 1 :date -1)]
+     ["messages" (array-map :username 1 :date -1)]
+     ["news" (array-map :date -1)]
+     ["users" (array-map :username 1)]
+     ["users" (array-map :username 1)
+      (assoc case-insensitive-index-opts :name "username_ci_1")]
+     ["users" (array-map :email 1)]
+     ["users" (array-map :email 1)
+      (assoc case-insensitive-index-opts :name "email_ci_1")]
+     ["users" (array-map :isadmin 1)]
+     ["users" (array-map :ismoderator 1)]
+     ["users" (array-map :special 1)]
+     ["users" (array-map :resetPasswordToken 1)]]))
+
 (defn create-indexes
-  "Create case insensitive indexes on `username` and `email` in the \"users\" collection.
+  "Create indexes for queries in our codebase.
 
   `create-indexes` can safely be executed multiple times, as long as the
   existing indexes don't conflict with the ones created here."
   []
   (webdb/connect)
   (try
-    (let [case-insensitive-index-opts {:collation {:locale "en" :strength (int 2)}}]
-      (mc/create-index db "users" (array-map :username 1) case-insensitive-index-opts)
-      (mc/create-index db "users" (array-map :email 1) case-insensitive-index-opts)
-      (println "Indexes successfully created.\n")
-      (println "All current indexes on \"users\":")
-      (clojure.pprint/pprint (mc/indexes-on db "users")))
+    (do
+      (doseq [index-args indexes]
+        (apply mc/create-index db index-args))
+      (println "Indexes successfully created."))
     (catch Exception e (do
                          (println "Create indexes failed" (.getMessage e))
                          (.printStackTrace e)))